textseg: Update tables for Unicode 12.0.0
As of Unicode 11 the grapheme cluster definition is derived from a
property defined in the emoji data, so we need to introduce a table from
there too now in addition to the table for the text segmentation
specification.

This also includes an updated table of grapheme cluster segmentation tests
derived from the Unicode character database version 12.0.0.