Module is now github.com/apparentlymart/go-textseg/v9
In order to support new versions of Unicode without imposing those changes
on existing callers expecting prior versions, this repository is switching
to a new versioning model where each major version of Unicode we support
gets a separate major version with the corresponding number.

This "v9" is therefore just a rebadge of the previous v1 release in order
to establish the convention. Subsequent Unicode major releases will have
their own major releases of this repository, so that they are all
simultaneously importable into the same program and so that it's clearer
in a downstream project's go.mod which version of Unicode they are using
the rules from.