autoversion: Wrapper that selects Unicode version automatically
The main functionality for this repository is Unicode version specific and
we have a separate module import path for each so that callers can select
the specific versions they need.

This new "autoversion" module is here to make life a little easier for
callers who just want to match the text segmentation algorithm/data with
the Unicode version being used in the Go standard library, so that unicode
support is consistent throughout the program.

It uses a compile-time switch based on the current Go version to select
the appropriate textseg major version. It's based on Go version rather
than Unicode version so that it can be a compile-time rather than a
runtime decision, or else we'd end up with the tables for all Unicode
versions linked into every program using this package. Unfortunately this
means that this package will need maintenance for each new version of Go
that's released, but hopefully that small annoyance here leads to less
annoyance for downstreams who can just upgrade this package to gain
support for new Go/Unicode versions.