autoversion: Wrapper that selects Unicode version automatically
The main functionality for this repository is Unicode version specific and
we have a separate module import path for each so that callers can select
the specific versions they need.

This new "autoversion" module is here to make life a little easier for
callers who just want to match the text segmentation algorithm/data with
the Unicode version being used in the Go standard library, so that unicode
support is consistent throughout the program.

It uses a runtime switch on unicode.Version to select one of the major
versions.