textseg: Update tables for Unicode 13.
Unicode 13's version of UAX #29 still uses the same segmentation algorithm
as Unicode 12, but has some changes to the character class tables that
the algorithm is given in terms of.

The test tables haven't been updated in Unicode 13 and are still passing,
and so the new tables seem to be broadly compatible with the Unicode 12
behavior and unlikely to cause significant problems for applications
previously built for Unicode 12's rules.